fixed the sales update and added global exception

+++ src/main/java/com/sparksupport/productsales/service/SalesService.java
@@ -1,15 +1,19 @@
 package com.sparksupport.productsales.service;
 
 import com.fasterxml.uuid.Generators;
+import com.sparksupport.productsales.dto.ProductDTO;
 import com.sparksupport.productsales.dto.ResponseDTO;
 import com.sparksupport.productsales.dto.SalesDTO;
+import com.sparksupport.productsales.exception.ResourceNotFoundException;
 import com.sparksupport.productsales.repository.SalesRepository;
 import lombok.extern.slf4j.Slf4j;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Value;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 
+import java.util.ArrayList;
 import java.util.List;
 
 @Slf4j
@@ -53,7 +57,11 @@ public class SalesService {
     public ResponseDTO getAllSales() {
         try {
             log.info("Trying to get all sales");
-            return new ResponseDTO(salesRepository.getAllSales(), 200, true, null);
+            List<SalesDTO> salesDTOList = salesRepository.getAllSales();
+            if (salesDTOList == null) {
+                throw new Exception("Sales is empty");
+            }
+            return new ResponseDTO(salesDTOList, HttpStatus.OK.value(), true, null);
         } catch (Exception e) {
             log.error("Error occurred while getting all sales");
             return new ResponseDTO(null, errorCode, true, e.getMessage());
@@ -63,56 +71,80 @@ public class SalesService {
     public ResponseDTO getSalesById(String id) {
         try {
             log.info("Trying to get sales by id : {}", id);
-            return new ResponseDTO(salesRepository.getSalesById(id), 200, true, null);
+            SalesDTO salesDTO = salesRepository.getSalesById(id);
+            if (salesDTO == null) {
+                throw new ResourceNotFoundException("No matching record found for the id " + id);
+            }
+            return new ResponseDTO(salesDTO, HttpStatus.OK.value(), true, null);
         } catch (Exception e) {
             log.error("Error occurred while getting sales by id : {}", id);
             return new ResponseDTO(null, errorCode, true, e.getMessage());
         }
     }
 
-    public ResponseDTO updateSalesById(String salesId, SalesDTO salesDTO) {
+    public ResponseDTO updateSalesById(String salesId, SalesDTO updatedSalesDTO) {
         try {
+
+
             log.info("Trying to update sales by id : {}", salesId);
-            String productId = salesRepository.getSalesById(salesId).getProductId();
-            log.info("Product id : {}", productId);
+            SalesDTO oldSalesDTO = salesRepository.getSalesById(salesId);
+            if (oldSalesDTO == null) {
+                throw new ResourceNotFoundException("No matching record found for the id " + salesId);
+            }
+            log.info("Product id : {}", oldSalesDTO.getProductId());
 
-            int productStock = productService.getAvailableProductQuantity(productId);
-            log.info("Product stock available : {}", productStock);
+            //making changes for returning the dto once updated.
+            updatedSalesDTO.setId(salesId);
+            updatedSalesDTO.setProductId(oldSalesDTO.getProductId());
 
-            if (productStock >= salesDTO.getQuantity()) {
-                salesRepository.updateSalesById(salesDTO.getQuantity(), salesDTO.getSalesDate(), salesId);
-                productService.updateProductQuantity(productId, salesDTO.getQuantity());
+            int productInStock = productService.getAvailableProductQuantity(oldSalesDTO.getProductId());
+            log.info("Product stock available : {}", productInStock);
+
+            int productInStockUpdated = productInStock + oldSalesDTO.getQuantity();
+            if (productInStockUpdated >= updatedSalesDTO.getQuantity()) {
+                salesRepository.updateSalesById(updatedSalesDTO.getQuantity(), updatedSalesDTO.getSalesDate(), salesId);
+                productInStockUpdated -= updatedSalesDTO.getQuantity();
+                productService.updateProductQuantity(oldSalesDTO.getProductId(), productInStockUpdated);
             } else {
-                log.error("Error occurred while updating sales the stock quantity : {} and trying to sell : {}", productStock, salesDTO.getQuantity());
-                return new ResponseDTO(null, errorCode, false, "Sales quantity is more than the stock quantity.");
+                log.error("Error occurred while updating sales the stock quantity : {} and trying to sell : {}", productInStockUpdated, updatedSalesDTO.getQuantity());
+                throw new Exception("Sales quantity is more than the stock quantity.");
             }
         } catch (Exception e) {
             log.error("Error occurred while updating sales by id:{}", salesId);
             return new ResponseDTO(null, errorCode, false, e.getMessage());
         }
-        return new ResponseDTO("Sales updated successfully!", 200, true, null);
+        return new ResponseDTO(updatedSalesDTO, 200, true, null);
 
     }
 
     public ResponseDTO insertSalesByProductId(String productId, List<SalesDTO> salesDTOList) {
         try {
-            int productStock = productService.getAvailableProductQuantity(productId);
+            List<SalesDTO> insertedSalesDTO = new ArrayList<>();
+            ProductDTO productDTO = (ProductDTO) productService.getProductById(productId).getData();
+            if (productDTO == null) {
+                throw new ResourceNotFoundException("No matching record found for the product id " + productId);
+            }
+            int productInStock = productDTO.getQuantity();
             for (SalesDTO salesDTO : salesDTOList) {
                 try {
                     log.info("Trying to insert sales : {}", salesDTO.toString());
                     salesDTO.setId(Generators.timeBasedGenerator().generate().toString());
-                    if (productStock > salesDTO.getQuantity()) {
+                    if (productInStock > salesDTO.getQuantity()) {
                         salesRepository.insertSales(salesDTO.getId(), salesDTO.getQuantity(), salesDTO.getSalesDate(), productId);
-                        productService.updateProductQuantity(productId, salesDTO.getQuantity());
-                        productStock -= salesDTO.getQuantity();
+                        productInStock -= salesDTO.getQuantity();
+                        productService.updateProductQuantity(productId, productInStock);
+                        insertedSalesDTO.add(salesDTO);
                     } else {
-                        log.error("Error occurred while inserting sales the stock quantity : {} and trying to sell : {}", productStock, salesDTO.getQuantity());
+                        log.error("Error occurred while inserting sales the stock quantity : {} and trying to sell : {}", productInStock, salesDTO.getQuantity());
                     }
                 } catch (Exception e) {
                     log.error("Error occurred while inserting sales: {}", salesDTO.toString());
                 }
             }
-            return new ResponseDTO("Sales inserted successfully!", 200, true, null);
+            if (insertedSalesDTO == null) {
+                throw new Exception("No record has been inserted.");
+            }
+            return new ResponseDTO(insertedSalesDTO, 200, true, null);
 
         } catch (Exception e) {
             log.error("Error occurred while inserting sales...");
@@ -123,6 +155,10 @@ public class SalesService {
     public ResponseDTO deleteSalesById(String id) {
         try {
             log.info("Trying to delete sales by id : {}", id);
+            SalesDTO SalesDTO = salesRepository.getSalesById(id);
+            if (SalesDTO == null) {
+                throw new ResourceNotFoundException("No matching record found for the id " + id);
+            }
             salesRepository.deleteSalesById(id);
         } catch (Exception e) {
             log.error("Error occurred while deleting sales by id:{}", id);
@@ -133,6 +169,15 @@ public class SalesService {
     }
 
     public List<String> getSalesByProductId(String productId) {
-        return salesRepository.getSalesByProductId(productId);
+        ProductDTO productDTO = (ProductDTO) productService.getProductById(productId).getData();
+        if(productDTO == null){
+            throw new ResourceNotFoundException("No matching product record found for the id : "+productId);
+        }
+        List<String> salesId = salesRepository.getSalesByProductId(productId);
+        if(salesId == null){
+            throw new ResourceNotFoundException("No matching sales record found for the id : "+salesId);
+        }
+
+        return salesId;
     }
 }
